refactor : edit all code from slug to uuid
